requirements.txt
aiohttp==3.7.3
APScheduler==3.6.3
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cryptography==3.3.1
decorator==4.4.2
idna==2.10
multidict==5.1.0
pycparser==2.20
python-box==5.2.0
python-telegram-bot==13.1
pytz==2020.4
requests==2.25.1
six==1.15.0
tornado==6.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.2
voluptuous==0.12.1
yandex-weather-api==0.2.4
yarl==1.6.3
